# ARROW-10654: [Rust] Specialize parsing of floats / bools in CSV Reader