Internal rust float parser is known to be slow.

This change allows to have specialized implementations rather than relying on FromStr::parse.

Also avoids calling `to_lowercase` for booleans.

Would be nice to benchmark this.