@Dandandan 2020-11-19
Some benchmark/context of string -> f64 is here https://github.com/Alexhuszagh/rust-lexical/