@Dandandan 2020-11-19
Did some benchmarking on this. Seems like a small win.

Master:
```
Running benchmarks with the following options: Opt { debug: false, iterations: 10, concurrency: 1, batch_size: 4096, path: "./yellow_tripdata_2020-01.csv", file_format: "csv" }

Query 'fare_amt_by_passenger' iteration 0 took 4114 ms
Query 'fare_amt_by_passenger' iteration 1 took 4087 ms
Query 'fare_amt_by_passenger' iteration 2 took 4094 ms
Query 'fare_amt_by_passenger' iteration 3 took 4118 ms
Query 'fare_amt_by_passenger' iteration 4 took 4091 ms
Query 'fare_amt_by_passenger' iteration 5 took 4099 ms
Query 'fare_amt_by_passenger' iteration 6 took 4115 ms
Query 'fare_amt_by_passenger' iteration 7 took 4129 ms
Query 'fare_amt_by_passenger' iteration 8 took 4105 ms
Query 'fare_amt_by_passenger' iteration 9 took 4095 ms
```

This version:
```
Running benchmarks with the following options: Opt { debug: false, iterations: 10, concurrency: 1, batch_size: 4096, path: "./yellow_tripdata_2020-01.csv", file_format: "csv" }
Executing 'fare_amt_by_passenger'
Query 'fare_amt_by_passenger' iteration 0 took 3985 ms
Query 'fare_amt_by_passenger' iteration 1 took 3954 ms
Query 'fare_amt_by_passenger' iteration 2 took 3961 ms
Query 'fare_amt_by_passenger' iteration 3 took 3959 ms
Query 'fare_amt_by_passenger' iteration 4 took 3963 ms
Query 'fare_amt_by_passenger' iteration 5 took 3963 ms
Query 'fare_amt_by_passenger' iteration 6 took 3958 ms
Query 'fare_amt_by_passenger' iteration 7 took 3959 ms
Query 'fare_amt_by_passenger' iteration 8 took 3971 ms
Query 'fare_amt_by_passenger' iteration 9 took 3977 ms
```